app/requirements.txt
# GUI
PySide6>=6.8

# Rendering
Pillow>=10.3.0
reportlab>=4.0.9

# Packaging (opsional, untuk build .app)
PyInstaller>=6.4

